First version of a new repositories section
This section uses the repositories present in the old files list, and
uses those directories to change to them and reload.

A nice use case is for people who open Vim in the home directory, and
want to jump to one of the recently used repositories. First they change
to the repository, and then they use the list of recently opened files
in that directory.